requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
python-jose[cryptography]
python-dotenv
openai
pydantic
bcrypt
python-multipart
email-validator
